+++ src/main/kotlin/com/ravit/stash/actuator/MongoReplicaSetDetails.kt
@@ -1,4 +1,4 @@
-package com.ravit.stash.health
+package com.ravit.stash.actuator
 
 data class MongoReplicaSetDetails(
     val nodes: List<MongoNodeHealth>,