+++ src/main/kotlin/com/ravit/stash/controller/externals/project/request/ProjectExternalRequest.kt
@@ -1,5 +1,6 @@
 package com.ravit.stash.controller.externals.project.request
 
+import com.ravit.stash.controller.externals.shared.request.PeriodExternalRequest
 import com.ravit.stash.shared.code.CompanyType
 
 data class ProjectExternalRequest(
@@ -7,7 +8,7 @@ data class ProjectExternalRequest(
     val company: CompanyType,
     val name: String,
     val summary: String,
-    val period: ProjectPeriodExternalRequest,
+    val period: PeriodExternalRequest?,
     val techStack: List<String> = emptyList(),
     val achievements: List<String> = emptyList(),
     val teamSize: Int? = null,