+++ src/main/kotlin/com/ravit/stash/controller/externals/project/request/ProjectPeriodExternalRequest.kt
@@ -1,8 +1,0 @@
-package com.ravit.stash.controller.externals.project.request
-
-import java.time.LocalDate
-
-data class ProjectPeriodExternalRequest(
-    val startedAt: LocalDate,
-    val endedAt: LocalDate?,
-)