+++ src/main/kotlin/com/ravit/stash/domain/task/document/Task.kt
@@ -1,26 +1,32 @@
 package com.ravit.stash.domain.task.document
 
 import com.ravit.stash.shared.code.TaskType
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.Id
-import org.springframework.data.annotation.LastModifiedDate
+import com.ravit.stash.shared.document.BaseDocument
+import com.ravit.stash.shared.document.Period
+import org.springframework.data.mongodb.core.index.Indexed
 import org.springframework.data.mongodb.core.mapping.Document
-import java.time.LocalDateTime
 
 @Document(collection = "tasks")
-data class Task(
-    @Id
-    val id: String,
+class Task(
+    @Indexed
     val projectId: String,
+    @Indexed
     val type: TaskType,
     val title: String,
     val description: String? = null,
-    val period: TaskPeriod? = null,
+    val period: Period? = null,
     val workingDays: Int? = null,
     val details: TaskDetails? = null,
-    val keywords: List<String> = emptyList(),
-    @CreatedDate
-    val createdAt: LocalDateTime = LocalDateTime.now(),
-    @LastModifiedDate
-    val updatedAt: LocalDateTime = LocalDateTime.now(),
+    var keywords: List<String> = emptyList(),
+) : BaseDocument() {
+    fun updateKeywords(newKeywords: List<String>) =
+        apply {
+            keywords = newKeywords
+        }
+}
+
+data class TaskDetails(
+    val background: String? = null,
+    val solution: String? = null,
+    val impact: String? = null,
 )