+++ src/main/kotlin/com/ravit/stash/domain/task/document/TaskPeriod.kt
@@ -1,8 +1,0 @@
-package com.ravit.stash.domain.task.document
-
-import java.time.LocalDate
-
-data class TaskPeriod(
-    val startedAt: LocalDate?,
-    val completedAt: LocalDate?,
-)