+++ src/main/kotlin/com/ravit/stash/shared/document/BaseDocument.kt
@@ -1,0 +1,21 @@
+package com.ravit.stash.shared.document
+
+import org.bson.types.ObjectId
+import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.Id
+import org.springframework.data.annotation.LastModifiedDate
+import java.time.LocalDateTime
+
+abstract class BaseDocument {
+    @Id
+    var id: ObjectId? = null
+        private set
+
+    @CreatedDate
+    var createdAt: LocalDateTime? = null
+        private set
+
+    @LastModifiedDate
+    var updatedAt: LocalDateTime? = null
+        private set
+}